# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kaas-ui",
    "description": "A read-only, multi-cluster Kafka UI. Every data route is a GET; there is no mutating endpoint in this document because there is none in the router.",
    "license": {
      "name": "Apache-2.0"
    },
    "version": "0.9.0"
  },
  "paths": {
...
    "/api/environments/{env}/clusters/{id}": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "`GET /api/clusters/{id}`",
        "description": "The card, the broker list, and `DescribeCluster` where the cluster answers\nit. Where it does not — which is one of the two development clusters today\n— the description is absent and the reason travels in `errors`. The page\nrenders, with a note.",
        "operationId": "detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cluster detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ClusterDetail"
                }
              }
            }
          },
          "404": {
            "description": "No such cluster"
          },
          "503": {
            "description": "Configured but not connected"
          }
        }
      }
    },
    "/api/environments/{env}/clusters/{id}/brokers": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "`GET /api/clusters/{id}/brokers`",
        "operationId": "brokers",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Brokers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Broker"
                }
              }
            }
          }
        }
      }
    },
    "/api/environments/{env}/clusters/{id}/brokers/{node}/log-dirs": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "`GET /api/clusters/{id}/brokers/{node}/log-dirs`",
        "description": "One of the four RPCs that go to a *specific* broker. A broker that is down\nyields an error for that node, not for the call — which is why this is\nper-node rather than a fan-out that one dead broker can blank.",
        "operationId": "log_dirs",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "node",
            "in": "path",
            "description": "Broker node id",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Log directories",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_LogDirDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/environments/{env}/clusters/{id}/capabilities": {
      "get": {
        "tags": [
          "capabilities"
        ],
        "summary": "`GET /api/clusters/{id}/capabilities`",
        "description": "kaas-lib's version table is **per connection**, deliberately: brokers\nmid-rolling-upgrade genuinely disagree, and a cluster-wide table would be\nwrong during exactly the window when being right matters. There is no\n`cluster.capabilities()` to project from, and fabricating one by using\nwhichever connection answered gives a UI whose tabs flicker.\n\nSo the table is read from an **explicitly named** broker — the lowest node\nid unless `?broker=` says otherwise — and `source` says which. The UI\nrenders that as \"as reported by broker 1\".",
        "operationId": "capabilities",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Cluster id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "broker",
            "in": "query",
            "description": "Read the table from this broker",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "What this cluster can be asked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Capabilities"
                }
              }
            }
          },
          "404": {
            "description": "No such cluster"
          },
          "503": {
            "description": "Configured but not connected"
          }
        }
      }
    },
...
    "/api/environments/{env}/schema-registries/{registry}/subjects/{subject}/versions": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "`GET /api/environments/{env}/schema-registries/{registry}/subjects/{subject}/versions`",
        "operationId": "versions",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "description": "Environment id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "registry",
            "in": "path",
            "description": "Schema registry id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subject",
            "in": "path",
            "description": "Subject name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Every registered version of a subject",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Broker": {
        "type": "object",
        "description": "A broker, as the snapshot reports it — optionally enriched.",
        "required": [
          "nodeId",
          "host",
          "port",
          "isController",
          "leaderPartitionCount",
          "replicaPartitionCount"
        ],
        "properties": {
          "host": {
            "type": "string",
            "description": "Advertised host."
          },
          "isController": {
            "type": "boolean",
            "description": "Whether this broker is the controller."
          },
          "isFenced": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the controller has fenced it.\n\n`None` on a cluster that does not implement `DescribeCluster`: absent,\nnot false. Rendering an unknown as \"healthy\" is the one thing a fleet\ndashboard must never do."
          },
          "leaderPartitionCount": {
            "type": "integer",
            "description": "Partitions this broker leads.",
            "minimum": 0
          },
          "nodeId": {
            "type": "integer",
            "format": "int32",
            "description": "Node id."
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "description": "Advertised port."
          },
          "rack": {
            "type": [
              "string",
              "null"
            ],
            "description": "Rack, when the broker declares one."
          },
          "replicaPartitionCount": {
            "type": "integer",
            "description": "Partition replicas hosted here.",
            "minimum": 0
          }
        }
      },
      "Capabilities": {
        "type": "object",
        "description": "The capability answer.",
        "required": [
          "features",
          "source",
          "apiKeys",
          "brokerAheadCount"
        ],
        "properties": {
          "apiKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyEntry"
            },
            "description": "The whole table, for the api-versions panel."
          },
          "brokerAheadCount": {
            "type": "integer",
            "description": "How many keys the broker advertises that this build cannot encode.\n\nNot a failure — a fact about a broker newer than the codec. Two of\nthese on the Strimzi cluster today cannot even be named.",
            "minimum": 0
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeatureEntry"
            },
            "description": "Feature name to state."
          },
          "source": {
            "$ref": "#/components/schemas/CapabilitySource",
            "description": "Which broker answered."
          }
        }
      },
...
      "ClusterDescriptionDto": {
        "type": "object",
        "description": "`DescribeCluster`, where the cluster implements it.",
        "required": [
          "clusterId"
        ],
        "properties": {
          "clusterId": {
            "type": "string",
            "description": "The broker-reported cluster id."
          },
          "controllerId": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The authoritative controller."
          }
        }
      },
...
      "Envelope_Broker": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "A broker, as the snapshot reports it — optionally enriched.",
              "required": [
                "nodeId",
                "host",
                "port",
                "isController",
                "leaderPartitionCount",
                "replicaPartitionCount"
              ],
              "properties": {
                "host": {
                  "type": "string",
                  "description": "Advertised host."
                },
                "isController": {
                  "type": "boolean",
                  "description": "Whether this broker is the controller."
                },
                "isFenced": {
                  "type": [
                    "boolean",
                    "null"
                  ],
                  "description": "Whether the controller has fenced it.\n\n`None` on a cluster that does not implement `DescribeCluster`: absent,\nnot false. Rendering an unknown as \"healthy\" is the one thing a fleet\ndashboard must never do."
                },
                "leaderPartitionCount": {
                  "type": "integer",
                  "description": "Partitions this broker leads.",
                  "minimum": 0
                },
                "nodeId": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Node id."
                },
                "port": {
                  "type": "integer",
                  "format": "int32",
                  "description": "Advertised port."
                },
                "rack": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "Rack, when the broker declares one."
                },
                "replicaPartitionCount": {
                  "type": "integer",
                  "description": "Partition replicas hosted here.",
                  "minimum": 0
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
      "Envelope_ClusterDetail": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "The cluster detail page.\n\nThe card, plus the two things a detail page adds: the broker list, and\n`DescribeCluster` **where the cluster implements it**. Where it does not,\n`description` is `None` and the failure travels in the envelope's `errors`\nas a named resource — the page renders the snapshot data with a note, not\nan error.",
              "required": [
                "cluster",
                "brokers"
              ],
              "properties": {
                "brokers": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Broker"
                  },
                  "description": "Brokers, from the snapshot, enriched where possible."
                },
                "cluster": {
                  "$ref": "#/components/schemas/ClusterCard",
                  "description": "Everything the fleet card shows."
                },
                "description": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/ClusterDescriptionDto",
                      "description": "`DescribeCluster`, where it answered."
                    }
                  ]
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
      "Envelope_LogDirDto": {
        "type": "object",
        "description": "What every data endpoint returns.",
        "required": [
          "items",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "The resources that did not. Non-empty is still `200 OK`."
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "One log directory on one broker.",
              "required": [
                "path",
                "replicas"
              ],
              "properties": {
                "error": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "description": "The directory's own error, if it has one."
                },
                "path": {
                  "type": "string",
                  "description": "Broker-local path."
                },
                "replicas": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogDirReplicaDto"
                  },
                  "description": "Replicas stored here."
                },
                "totalBytes": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "Total capacity, where the broker reports it."
                },
                "usableBytes": {
                  "type": [
                    "integer",
                    "null"
                  ],
                  "format": "int64",
                  "description": "Free capacity."
                }
              }
            },
            "description": "The resources that answered."
          },
          "snapshotAgeMs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Age of the metadata snapshot the answer was built from, where one was\ninvolved. This is what makes \"as of 4 seconds ago\" honest.",
            "minimum": 0
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "description": "How many resources matched before paging or truncation.\n\nPresent only where the endpoint pages. Filtering and sorting happen on\nthe server — a fleet with a five-thousand-topic cluster in it is real —\nso the browser needs to be told what it is looking at a window into.",
            "minimum": 0
          }
        }
      },
...
      "ResourceError": {
        "type": "object",
        "description": "One resource's failure, named.",
        "required": [
          "resource",
          "kind",
          "message",
          "retriable"
        ],
        "properties": {
          "code": {
            "type": [
              "string",
              "null"
            ],
            "description": "The broker's error code name, when this build has one for it."
          },
          "codeNumber": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "The error code's number.\n\nPresent even when `code` is null: against a broker newer than the codec\n`ErrorCode::Unknown(i16)` is all there is, and the number is the only\nsearchable thing."
          },
          "kind": {
            "$ref": "#/components/schemas/ErrorKind",
            "description": "The taxonomy."
          },
          "message": {
            "type": "string",
            "description": "The error, rendered."
          },
          "resource": {
            "type": "string",
            "description": "Which resource — a topic name, a broker id, a group id."
          },
          "retriable": {
            "type": "boolean",
            "description": "Whether retrying this one resource is worth offering."
          },
          "unsupportedApi": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/UnsupportedApiDetail",
                "description": "Both version ranges, when the failure was an unsupported api."
              }
            ]
          }
        }
      },
...
      "SubjectDetail": {
        "type": "object",
        "description": "The versions of one subject, newest last.",
        "required": [
          "subject",
          "versions",
          "errors"
        ],
        "properties": {
          "compatibility": {
            "type": [
              "string",
              "null"
            ],
            "description": "The compatibility mode, where the registry reports one."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceError"
            },
            "description": "Versions the registry listed and would not hand over, with why.\n\nPartial failure is a result: a subject with thirty versions of which\none is unreadable renders twenty-nine and names the thirtieth."
          },
          "registry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RegistryCard",
                "description": "The registry answering."
              }
            ]
          },
          "subject": {
            "type": "string",
            "description": "The subject that was asked for."
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubjectSchema"
            },
            "description": "Every registered version, oldest first, with its text."
          }
        }
      },
...
    }
  },
...
}